Extend range of comparision counter
- Change datatype: int -> unsigned long long (ull)

+++ insertion_sort.cpp
@@ -1,6 +1,6 @@
 #include "sorting_algorithm.h"
 
-void insertionSort(vector<type>& v, int& comparision_count) {
+void insertionSort(vector<type>& v, ull& comparision_count) {
 	int size = v.size();
 
 	for (int i = 1; ++comparision_count && i < size; i++) {
